Fix more clang-tidy issues

+++ src/metaparser.cpp
@@ -99,7 +99,7 @@ MetadataParseException::MetadataParseException(const std::string& what_arg)
 namespace details
 {
 
-long msf_to_frames(const int m, const int s, const int f)
+int64_t msf_to_frames(const int m, const int s, const int f)
 {
 	if (m < 0 || m > 99 || s < 0 || s >= 60 || f < 0 || f >= 75) {
 		return -1;
@@ -109,7 +109,7 @@ long msf_to_frames(const int m, const int s, const int f)
 }
 
 
-void frames_to_msf(long frames, int* m, int* s, int* f)
+void frames_to_msf(int64_t frames, int64_t* m, int64_t* s, int64_t* f)
 {
 	*f = frames % 75;
 	frames /= 75;
